+++ src/main/java/org/snomed/snowstorm/rest/pojo/ItemsPage.java
@@ -31,7 +31,7 @@ public class ItemsPage<T> {
 		this.items = page.getContent();
 		this.limit = page.getSize();
 		this.total = page.getTotalElements();
-		this.offset = page.getNumber() * page.getSize();
+		this.offset = page.getNumber() * (long)page.getSize();
 	}
 
 	@JsonView(View.Component.class)
